chore: remove file configuration
Mutator file configuration is a deprecated feature since it's never
been used nor there is a need for it to be used since the tool is too
slow to be used on the whole project.
