+++ move-mutation-test/src/cli.rs
@@ -8,12 +8,10 @@ use clap::Parser;
 use move_model::metadata::LanguageVersion;
 use move_mutator::cli::{FunctionFilter, ModuleFilter};
 use move_package::CompilerConfig;
-use serde::{Deserialize, Serialize};
 use std::path::PathBuf;
 
 /// Command line options for mutation test tool.
-#[derive(Parser, Default, Debug, Clone, Deserialize, Serialize)]
-#[serde(default, deny_unknown_fields)]
+#[derive(Parser, Default, Debug, Clone)]
 pub struct CLIOptions {
     /// Work only over specified modules.
     #[clap(
@@ -32,10 +30,6 @@ pub struct CLIOptions {
         conflicts_with = "use_generated_mutants"
     )]
     pub mutate_functions: FunctionFilter,
-
-    /// Optional configuration file for mutator tool.
-    #[clap(long, value_parser, conflicts_with = "use_generated_mutants")]
-    pub mutator_conf: Option<PathBuf>,
 
     /// Save report to a JSON file.
     #[clap(long, value_parser)]
@@ -59,26 +53,12 @@ pub fn create_mutator_options(
     move_mutator::cli::CLIOptions {
         mutate_functions: options.mutate_functions.clone(),
         mutate_modules: options.mutate_modules.clone(),
-        configuration_file: options.mutator_conf.clone(),
         downsampling_ratio_percentage: options.downsampling_ratio_percentage,
         apply_coverage,
         // To run tests, compilation must succeed
         verify_mutants: true,
         ..Default::default()
     }
-}
-
-/// This function checks if the mutator output path is provided in the configuration file.
-///
-/// We don't need to check if the mutator output path is provided in the options as they were created
-/// from the mutation-test options which does not allow setting it.
-#[must_use]
-pub fn check_mutator_output_path(options: &move_mutator::cli::CLIOptions) -> Option<PathBuf> {
-    options
-        .configuration_file
-        .as_ref()
-        .and_then(|conf| move_mutator::configuration::Configuration::from_file(conf).ok())
-        .and_then(|c| c.project.out_mutant_dir)
 }
 
 /// The configuration options for running the tests.
@@ -144,14 +124,12 @@ fn get_bytecode_version(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::{fs, path::PathBuf};
 
     #[test]
     fn cli_options_starts_empty() {
         let options = CLIOptions::default();
         assert_eq!(ModuleFilter::All, options.mutate_modules);
         assert_eq!(FunctionFilter::All, options.mutate_functions);
-        assert!(options.mutator_conf.is_none());
         assert!(options.output.is_none());
     }
 
@@ -163,45 +141,11 @@ mod tests {
                 "func1".to_string(),
                 "func2".to_string(),
             ]),
-            mutator_conf: Some(PathBuf::from("path/to/mutator/conf")),
             ..Default::default()
         };
 
         let mutator_options = create_mutator_options(&options, false);
 
         assert_eq!(mutator_options.mutate_modules, options.mutate_modules);
-        assert_eq!(mutator_options.configuration_file, options.mutator_conf);
-    }
-
-    #[test]
-    fn check_mutator_output_path_returns_none_when_no_conf() {
-        let options = move_mutator::cli::CLIOptions::default();
-        assert!(check_mutator_output_path(&options).is_none());
-    }
-
-    #[test]
-    fn check_mutator_output_path_returns_path_when_conf_exists() {
-        let json_content = r#"
-            {
-                "project": {
-                    "out_mutant_dir": "path/to/out_mutant_dir"
-                },
-                "project_path": "/path/to/project",
-                "individual": []
-            }
-        "#;
-
-        fs::write("test_mutator_conf.json", json_content).unwrap();
-
-        let options = move_mutator::cli::CLIOptions {
-            configuration_file: Some(PathBuf::from("test_mutator_conf.json")),
-            ..Default::default()
-        };
-
-        let path = check_mutator_output_path(&options);
-        fs::remove_file("test_mutator_conf.json").unwrap();
-
-        assert!(path.is_some());
-        assert_eq!(path.unwrap(), PathBuf::from("path/to/out_mutant_dir"));
     }
 }
