+++ move-mutation-test/tests/integration_tests.rs
@@ -33,7 +33,7 @@ fn test_run_mutation_test(path: &Path, expected_report: String) -> datatest_stab
         dump_state: false,
         filter: None,
         ignore_compile_warnings: false,
-        // TODO(rqnsom): maybe we could do set it to true, but it would require `aptos` command in
+        // TODO(rqnsom): maybe we could set it to true, but it would require `aptos` command in
         // the `build.rs` - using `process::Command` slowed down the execution a lot
         apply_coverage: false,
         gas_limit: 2000,
