+++ move-mutator/src/configuration.rs
@@ -3,31 +3,16 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{cli::CLIOptions, coverage::Coverage};
-use serde::{Deserialize, Serialize};
-use std::{
-    path::{Path, PathBuf},
-    str::FromStr,
-};
+use std::path::PathBuf;
 
-/// Configuration file type.
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
-pub enum FileType {
-    JSON,
-    TOML,
-}
 /// Mutator configuration for the Move project.
-#[derive(Debug, Serialize, Deserialize, Default)]
+#[derive(Debug, Default)]
 pub struct Configuration {
     /// Main project options. It's the same as the CLI options.
     pub project: CLIOptions,
     /// Path to the project.
     pub project_path: Option<PathBuf>,
-    /// Configuration for the mutation operators (project-wide).
-    pub mutation: Option<MutationConfig>,
-    /// Configuration for the individual files. (optional).
-    pub individual: Vec<FileConfiguration>,
     /// Coverage report where the optional unit test coverage data is stored.
-    #[serde(skip)]
     pub(crate) coverage: Coverage,
 }
 
@@ -38,303 +23,8 @@ impl Configuration {
         Self {
             project,
             project_path,
-            mutation: None,
-            individual: vec![],
             // Coverage is disabled by default.
             coverage: Coverage::default(),
         }
-    }
-
-    /// Recognizes the file type based on the file extension.
-    /// Currently supported file types are JSON and TOML.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the file type is not supported.
-    ///
-    /// # Returns
-    ///
-    /// * `FileType` - The recognized file type.
-    fn get_file_type(file_path: &Path) -> anyhow::Result<FileType> {
-        match file_path.extension().and_then(|s| s.to_str()) {
-            Some("json") => Ok(FileType::JSON),
-            Some("toml") => Ok(FileType::TOML),
-            _ => Err(anyhow::anyhow!("Unsupported file type")),
-        }
-    }
-
-    /// Reads configuration from the configuration file recognizing its type.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the file type is unsupported or the configuration has an invalid format.
-    ///
-    /// # Returns
-    ///
-    /// * `Configuration` - The configuration read from the file.
-    pub fn from_file(config_path: &Path) -> anyhow::Result<Configuration> {
-        let file_type = Configuration::get_file_type(config_path)?;
-        debug!("Reading configuration from file type: {:?}", file_type);
-        match file_type {
-            FileType::JSON => Configuration::from_json_file(config_path),
-            FileType::TOML => Configuration::from_toml_file(config_path),
-        }
-    }
-
-    /// Reads configuration from the TOML configuration file.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the file has an invalid format.
-    ///
-    /// # Returns
-    ///
-    /// * `Configuration` - The configuration read from the file.
-    pub fn from_toml_file(toml_file: &Path) -> anyhow::Result<Configuration> {
-        debug!("Reading configuration from TOML file: {:?}", toml_file);
-        let toml_source = std::fs::read_to_string(toml_file)?;
-        Ok(toml::from_str(toml_source.as_str())?)
-    }
-
-    /// Reads configuration from the JSON configuration source.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the file has an invalid format.
-    ///
-    /// # Returns
-    ///
-    /// * `Configuration` - The configuration read from the file.
-    pub fn from_json_file(json_file: &Path) -> anyhow::Result<Configuration> {
-        debug!("Reading configuration from JSON file: {:?}", json_file);
-        Ok(serde_json::from_str(&std::fs::read_to_string(json_file)?)?)
-    }
-
-    /// Returns the configuration for the given file path.
-    #[must_use]
-    pub fn get_file_configuration(&self, file_path: &Path) -> Option<&FileConfiguration> {
-        self.individual
-            .iter()
-            .find(|file_conf| file_conf.file == file_path)
-    }
-}
-
-/// Configuration of the mutation operators.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct MutationConfig {
-    /// Names of the mutation operators to use. If not provided, all operators will be used.
-    pub operators: Vec<String>,
-    /// Names of the mutation categories to be used.
-    pub categories: Vec<String>,
-}
-
-/// Configuration for the individual file.
-#[derive(Debug, Serialize, Deserialize)]
-#[allow(clippy::module_name_repetitions)]
-pub struct FileConfiguration {
-    /// The path to the Move source.
-    pub file: PathBuf,
-    /// Indicates if the mutants should be verified.
-    pub verify_mutants: bool,
-    /// Names of the mutation operators to use. If not provided, all operators will be used.
-    pub mutation_operators: Option<MutationConfig>,
-    /// Mutate only the functions with the given names.
-    pub mutate_functions: FunctionFilter,
-}
-
-/// Filter for the functions to mutate.
-#[derive(Default, Debug, Clone, Deserialize, Serialize, PartialEq)]
-pub enum FunctionFilter {
-    #[default]
-    All,
-    Selected(Vec<String>),
-}
-
-impl FromStr for FunctionFilter {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "all" => Ok(FunctionFilter::All),
-            _ => Ok(FunctionFilter::Selected(
-                s.split(&[';', '-', ',']).map(String::from).collect(),
-            )),
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::cli::ModuleFilter;
-    use std::{fs, path::Path};
-
-    #[test]
-    fn configuration_from_toml_file_loads_correctly() {
-        let toml_content = r#"
-            [project]
-            move_sources = ["/path/to/move/source"]
-            mutate_modules = {"Selected" = ["module1", "module2"]}
-            [mutation]
-            operators = ["operator1", "operator2"]
-            categories = ["category1", "category2"]
-            [[individual]]
-            file = "/path/to/file"
-            verify_mutants = true
-            mutate_functions = "All"
-        "#;
-        fs::write("test.toml", toml_content).unwrap();
-        let config = Configuration::from_toml_file(Path::new("test.toml")).unwrap();
-        fs::remove_file("test.toml").unwrap();
-        assert_eq!(
-            config.project.move_sources,
-            vec![Path::new("/path/to/move/source")]
-        );
-        assert_eq!(
-            config.project.mutate_modules,
-            ModuleFilter::Selected(vec!["module1".to_owned(), "module2".to_owned()])
-        );
-        assert_eq!(
-            config.mutation.unwrap().operators,
-            vec!["operator1", "operator2"]
-        );
-        assert_eq!(config.individual.len(), 1);
-        assert_eq!(config.individual[0].file, PathBuf::from("/path/to/file"));
-        assert!(config.individual[0].verify_mutants);
-        assert_eq!(config.individual[0].mutate_functions, FunctionFilter::All);
-    }
-
-    #[test]
-    fn configuration_from_non_existent_toml_file_fails() {
-        let result = Configuration::from_toml_file(Path::new("non_existent.toml"));
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn configuration_from_invalid_toml_file_fails() {
-        let toml_content = r#"
-            [project]
-            move_sources = "/path/to/move/source"
-        "#;
-        fs::write("test_invalid.toml", toml_content).unwrap();
-        let result = Configuration::from_toml_file(Path::new("test_invalid.toml"));
-        fs::remove_file("test_invalid.toml").unwrap();
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn configuration_from_json_file_loads_correctly() {
-        let json_content = r#"
-            {
-                "project": {
-                    "move_sources": ["/path/to/move/source"],
-                    "mutate_modules": "All",
-                    "out_mutant_dir": "/path/to/output",
-                    "verify_mutants": true,
-                    "no_overwrite": false,
-                    "configuration_file": "/path/to/configuration"
-                },
-                "project_path": "/path/to/project",
-                "mutation": {
-                    "operators": ["operator1", "operator2"],
-                    "categories": ["category1", "category2"]
-                },
-                "individual": [
-                    {
-                        "file": "/path/to/file",
-                        "verify_mutants": true,
-                        "mutation_operators": {
-                            "operators": ["operator3", "operator4"],
-                            "categories": ["category3", "category4"]
-                        },
-                        "mutate_functions": "All"
-                    }
-                ]
-            }
-        "#;
-        fs::write("test.json", json_content).unwrap();
-        let config = Configuration::from_json_file(Path::new("test.json")).unwrap();
-        fs::remove_file("test.json").unwrap();
-        assert_eq!(
-            config.project.move_sources,
-            vec![Path::new("/path/to/move/source")]
-        );
-        assert_eq!(config.project.mutate_modules, ModuleFilter::All);
-        assert_eq!(
-            config.project.out_mutant_dir,
-            Some(PathBuf::from("/path/to/output"))
-        );
-        assert!(config.project.verify_mutants);
-        assert!(!config.project.no_overwrite);
-        assert_eq!(
-            config.project.configuration_file.unwrap(),
-            Path::new("/path/to/configuration")
-        );
-        assert_eq!(config.project_path.unwrap(), Path::new("/path/to/project"));
-        assert_eq!(
-            config.mutation.unwrap().operators,
-            vec!["operator1", "operator2"]
-        );
-    }
-
-    #[test]
-    fn configuration_from_non_existent_json_file_fails() {
-        let result = Configuration::from_json_file(Path::new("non_existent.json"));
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn configuration_from_invalid_json_file_fails() {
-        let json_content = r#"
-            {
-                "project": {
-                    "move_sources": "/path/to/move/source"
-                }
-            }
-        "#;
-        fs::write("test_invalid.json", json_content).unwrap();
-        let result = Configuration::from_json_file(Path::new("test_invalid.json"));
-        fs::remove_file("test_invalid.json").unwrap();
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn recognizes_json_file_type_correctly() {
-        assert_eq!(
-            Configuration::get_file_type(Path::new("test.json")).unwrap(),
-            FileType::JSON
-        );
-    }
-
-    #[test]
-    fn recognizes_toml_file_type_correctly() {
-        assert_eq!(
-            Configuration::get_file_type(Path::new("test.toml")).unwrap(),
-            FileType::TOML
-        );
-    }
-
-    #[test]
-    fn configuration_from_file_fails_for_unknown_file_type() {
-        let result = Configuration::from_file(Path::new("test.unknown"));
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn get_file_configuration_returns_none_for_unknown_file() {
-        let file_path = PathBuf::from("/path/to/file");
-        let file_config = FileConfiguration {
-            file: file_path.clone(),
-            verify_mutants: true,
-            mutation_operators: None,
-            mutate_functions: FunctionFilter::All,
-        };
-        let config = Configuration {
-            individual: vec![file_config],
-            ..Default::default()
-        };
-
-        let result = config.get_file_configuration(&PathBuf::from("/unknown/path"));
-        assert!(result.is_none());
     }
 }
