+++ move-spec-test/src/cli.rs
@@ -4,12 +4,10 @@
 
 use clap::Parser;
 use move_mutator::cli::{FunctionFilter, ModuleFilter, PackagePathCheck};
-use serde::{Deserialize, Serialize};
 use std::path::PathBuf;
 
 /// Command line options for specification test tool.
-#[derive(Parser, Default, Debug, Clone, Deserialize, Serialize)]
-#[serde(default, deny_unknown_fields)]
+#[derive(Parser, Default, Debug, Clone)]
 pub struct CLIOptions {
     /// The paths to the Move sources.
     #[clap(long, value_parser)]
@@ -79,7 +77,6 @@ pub fn create_mutator_options(options: &CLIOptions) -> move_mutator::cli::CLIOpt
         move_sources: options.move_sources.clone(),
         mutate_modules: options.mutate_modules.clone(),
         mutate_functions: options.mutate_functions.clone(),
-        configuration_file: options.mutator_conf.clone(),
         verify_mutants: options.verify_mutants,
         downsampling_ratio_percentage: options.downsampling_ratio_percentage,
         ..Default::default()
@@ -100,21 +97,6 @@ pub fn generate_prover_options(options: &CLIOptions) -> anyhow::Result<move_prov
     };
 
     Ok(prover_conf)
-}
-
-/// This function checks if the mutator output path is provided in the configuration file.
-/// We don't need to check if the mutator output path is provided in the options as they were created
-/// from the spec-test options which does not allow setting it.
-#[must_use]
-pub fn check_mutator_output_path(options: &move_mutator::cli::CLIOptions) -> Option<PathBuf> {
-    if let Some(conf) = &options.configuration_file {
-        let c = move_mutator::configuration::Configuration::from_file(conf);
-        if let Ok(c) = c {
-            return c.project.out_mutant_dir;
-        }
-    };
-
-    None
 }
 
 #[cfg(test)]
@@ -149,40 +131,6 @@ mod tests {
         assert_eq!(mutator_options.move_sources, options.move_sources);
         assert_eq!(mutator_options.mutate_modules, options.mutate_modules);
         assert_eq!(mutator_options.mutate_functions, options.mutate_functions);
-        assert_eq!(mutator_options.configuration_file, options.mutator_conf);
-    }
-
-    #[test]
-    fn check_mutator_output_path_returns_none_when_no_conf() {
-        let options = move_mutator::cli::CLIOptions::default();
-        assert!(check_mutator_output_path(&options).is_none());
-    }
-
-    #[test]
-    fn check_mutator_output_path_returns_path_when_conf_exists() {
-        let json_content = r#"
-            {
-                "project": {
-                    "move_sources": ["/path/to/move/source"],
-                    "out_mutant_dir": "path/to/out_mutant_dir"
-                },
-                "project_path": "/path/to/project",
-                "individual": []
-            }
-        "#;
-
-        fs::write("test_mutator_conf.json", json_content).unwrap();
-
-        let options = move_mutator::cli::CLIOptions {
-            configuration_file: Some(PathBuf::from("test_mutator_conf.json")),
-            ..Default::default()
-        };
-
-        let path = check_mutator_output_path(&options);
-        fs::remove_file("test_mutator_conf.json").unwrap();
-
-        assert!(path.is_some());
-        assert_eq!(path.unwrap(), PathBuf::from("path/to/out_mutant_dir"));
     }
 
     #[test]
